requirement.txt
exception==0.1.0
imutils==0.4.6
numpy==1.14.3
opencv-python==3.4.1.15
Pillow==5.1.0
pytesseract==0.2.2
PyYAML==3.12
six==1.11.0
tesseract==0.1.3
virtualenv==16.0.0
XlsxWriter==1.0.5
